requirements.txt
torch
flask
torchvision
pillow
flask_cors